move all jsx files to components directory
