+++ challenge-react/package.json
@@ -10,7 +10,7 @@
     "server": "json-server --watch db.json --port 3001",
     "client": "webpack serve",
     "test": "jest",
-    "shallow": "yarn prettier -w ./src/*.{js,jsx}",
+    "shallow": "yarn prettier -w ./src/*.js",
     "deep": "yarn prettier -w ./src/**/*.{js,jsx}",
     "fmt": "yarn shallow && yarn deep"
   },
