+++ challenge-react/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import { App } from './App.jsx';
+import { App } from './components/App.jsx';
 import { actionTypes } from './actions';
 
 const store = createStore((state, action) => {
